bitkeeper revision 1.1159.17.14 (411b1b16kI8OhQTxCf1kbM8dV6LXcA)
prepare 2.6 blkfront for suspend/resume

+++ linux-2.6.7-xen-sparse/drivers/char/mem.c
@@ -252,7 +252,7 @@ static int mmap_mem(struct file * file, struct vm_area_struct * vma)
 	vma->vm_page_prot = pgprot_noncached(vma->vm_page_prot);
 	if (direct_remap_area_pages(vma->vm_mm, vma->vm_start, offset, 
 				vma->vm_end-vma->vm_start, vma->vm_page_prot,
-				(domid_t)file->private_data))
+				(domid_t)(unsigned long)file->private_data))
 		return -EAGAIN;
 	return 0;
 }
